// src/containers/tasks/enums.ts
export enum UsersActionTypes {
	ADD_LIST = 'ADD_LIST',
	UPDATE_LIST_NAME = 'UPDATE_LIST_NAME',
	SET_USER = 'SET_USER',
	SET_TASKS = 'SET_TASKS',
	SET_TASK = 'SET_TASK',
	ADD_USER_LIST_COLUMNS = 'ADD_USER_LIST_COLUMNS',
	DELETE_CATEGORY_COLUMN = 'DELETE_CATEGORY_COLUMN',
	SET_TASK_TITLE = 'SET_TASK_TITLE',
	SET_USERS = 'SET_USERS',
	REMOVE_LIST = 'REMOVE_LIST',
	ADD_DEFAULT_TASK = 'ADD_DEFAULT_TASK',
	DELETE_DEFAULT_TASK = 'DELETE_DEFAULT_TASK',
	UPDATE_DEFAULT_TASK = 'UPDATE_DEFAULT_TASK',
	CHANGE_DEFAULT_TASK_STATUS = 'CHANGE_DEFAULT_TASK_STATUS',
}


// export enum TasksActionTypes {
// 	SET_ACTIVE_DAY = 'SET_ACTIVE_DAY',
// 	SET_ACTIVE_CHARACTER = 'SET_ACTIVE_CHARACTER',
// }